Create and separate the jss files
