+++ src/App.js
@@ -18,7 +18,7 @@ import { getUserData, logout } from './redux/actions/userActions'
 const token = localStorage.FBIdToken;
 if (token) {
   const decodedToken = jwtDecode(token);
-  if (decodedToken.exp * 10000 < Date.now()) {
+  if (decodedToken.exp * 1000 < Date.now()) {
     store.dispatch(logout());
     window.location.href = '/'; 
   } else {
