+++ src/components/sentences/CreateSentence.js
@@ -10,7 +10,6 @@ import DialogContent from '@material-ui/core/DialogContent'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faPlusSquare } from '@fortawesome/free-solid-svg-icons'
 import styles from '../../styles/sentences/CreateSentenceStyles'
-import '../../styles/sentences/CreateSentenceForm.css'
 
 // Redux
 import { connect } from 'react-redux'
