+++ src/styles/sentences/CreateSentenceForm.css
@@ -3,6 +3,7 @@
   width: 100%;
   border: none;
   border-bottom: 1px solid grey;
+  margin-top: 1rem;
   height: 2.5rem;
   font-size: 0.95rem;
   outline: none;
