+++ src/styles/sentences/CreateSentenceStyles.js
@@ -2,16 +2,9 @@ const styles = {
   button: {
     color: "white"
   },
-  englishJapaneseInputForm: {
-    paddingRight: "none"
-  },
   addWord: {
     marginLeft: 25,
     marginTop: 25,
-  },
-  submitButton: {
-    marginTop: 5,
-    textAlign: 'center'
   }
 }
