+++ src/styles/sentences/UpdateSentenceStyles.js
