+++ src/styles/sentences/WordMeaningListStyles.js
@@ -1,0 +1,18 @@
+const styles = {
+  word: {
+    display: "flex",
+    alignItems: "center"
+  },
+  indices: {
+    width: "9%",
+    marginLeft: 15
+  },
+  english: {
+    fontSize: '1.1rem',
+    color: "blue",
+    width: "30%",
+    marginLeft: 15
+  },
+}
+
+export default styles
